move all tests into one file
